Further readability improvements
